Added Systems Story video
This is my attempt at embedding the "Systems Story" video into the SEE file. From what I understand from forums, the video will not appear in the document itself, but will link to the video in a browser. I am not sure where you want the video, Dominique, so feel free to move the code!