Add files via upload
FYI - none of the images in this iteration are formatted in terms of size because when I move it to Outlook the formatting doesn't keep anyways (I just take one corner and drag to make it smaller).  Adding unnecessary lines of code complicates other items sometimes when I add new lines of code so I left this out.