Update learning_objectives.rmd
Based on our discussion during the facilitation meeting, I tried to distill the several points down to a "core objective" with hopes that it would address the 'function' of each session and allow for additional 1-2 sub-points to be made that could be the 'topography' or more specifics about how we would achieve the core objective.  I started out by taking a look at half of them. If these look helpful / or seem like a useful way to approach this, please let me know and I can do the remaining half.