Update learning_objectives.rmd
I tried to distill each of these down to one "core objective" based on our discussion during the facilitation meeting about wanting to identify what was critical (the agent of change / indispensable) to each session, with the hope that it addresses the core 'function' of each session, and that additional sub points could be made with the 'topography' of how we plan to achieve the objective with more specifics. I've gone through half of them; if it looks like a useful approach / can move the project forward, please let me know and I can address the second half.