Add files via upload
RMD file of CPT Dropout and CDW Data Visualizations. I created a new file in the datascripts branch of the master as this is a new set of code in RMD form that I have been generating holding the 3 principles of coding in mind in order to be as transparent and reproducible as possible in my workflow: a) I tried to make it as human readable as possible, following an order, attempting to put assignment and transformation together, and trying to make sure nothing is hidden; b) I also tried to construct my script to make sure that everything needed for someone else to use it is ideally visible: specifically noted that a YAML header is present, my working directory/importing code is visible, my transforming and assigning is present (although I intentionally have put the numeric coaxing below my generation ofa  new merged dataframe. I wanted to think about it as creating my data source first and hten manipulating the datasource, rather than manipulating it prior to creating the data source - I welcome further discussion on this), as well as descriptives/visualization. As you will see, I generally did not get to the Analyses yet; c) I also attempted to annotate my rationale and description extensively.