Update learning_objectives.rmd
Debbie and I worked on LOs for Sessions 1-5. If there's agreement on our assumptions and approach, we can work on the rest in a similar vein. 
Our assumption is that the "official" LOs do not have to cover everything we will teach or want folks to learn; they need to serve 3 primary purposes: 
1. Be acceptable to EES for accreditation
2. Be able to yield reasonable evaluation results when the Learning system asks participants to rate their agreement with "Having participated in the session I am able to...(LO)"
3. Be appealing to potential participants, as they are part of how the learning opportunity is presented to them

We tried therefore to
1. Meet the minimum requirement (3 per session)
2. Not go over 3 without good reason, in order to keep expectations manageable for the length of sessions
3. Avoid jargon and advanced vocabulary
4. Avoid compound objectives
5. Repeat language from the user interface to reinforce learning, e.g. in describing the lessons of each main loop

We look forward to your thoughts on the above, as well as the LO edits proposed!