April 2018 Learning Objectives - All
These are the objectives for all sessions with the engineers' and facilitators' best thinking. They are ready for clinician thinking! We left specific questions in the document where we were unsure of appropriateness; but everyone should feel free to make other suggestions and edits as well. I think we may be able to track them best if they stay within the document - just ID your comments by color or initials and re-upload.

Stacey, I'm willing to take back that recommendation and try working in R if you can load the draft in as code and help us with a process for "conversation" on small or global points.

Thanks everyone!
Jane