Create comparatorclinics
Hi Savet,

Thank you for reviewing this code for me. I used most of the same code that you already cleaned and reviewed for me for the anti_join code.  In order to make sure that the CDW_minus_merge data we are looking at is only the data with dates that match the CPT dataset, I entered the following based on the timeframe, so the new code starts at Line 80:  filter(year_q == 2012Q2:2016Q3) %>%

For the second issue of making sure that we are looking at number of clinics and not number of observations, I believe I should be doing the analyses based on unique number of sta6a and sta3n combinations versus observations. To do this, I'm trying to generate a concatenated variable sta6asta3n. When I examine the frequencies of this I can tell the unique number of combinations. Is that the right way to think about this? 

Additionally however, I'm not very clear on how to run subsequent descriptors of other variables (e.g., total encounters) that is not a mean of each observation of total encounters, but instead is a mean of total encounters based on unique combinations of sta6a & sta3ns (i.e., clinics). i.e., I believe I still cannot do describe(cdw_minus_merge.df$totalencounters) for example, because that will still give me descriptive statistics on totalencounters based on observations and not unique sta6asta3n combinations. Can you please help me with this code?

Thank you!