Review and help on session 7 in-session exercise
Please review/edit the step-by-step guide for the exercise to create a base case run. I made a gif for the first 2 steps and copied Stacey's code to insert it into the guide (yay!) - but I wasn't sure where to save the gif; and the gif is fuzzy. Need some help on those things. Thanks!