Questions for Team PSD - Session 3 mutual learning activity
If the *MTL.how/menu* has already been completed in the in-session activity, we could show results of another teams' *MTL.HOW/MENU* completion. I also wondered about having the 9 participants draw a "role" out of a bowl so their discussion considers all points of view on a clinic team.