Review and help on session 7 in-session exercise (#135)
* Review and help on session 7 in-session exercise

Please review/edit the step-by-step guide for the exercise to create a base case run. I made a gif for the first 2 steps and copied Stacey's code to insert it into the guide (yay!) - but I wasn't sure where to save the gif; and the gif is fuzzy. Need some help on those things. Thanks!

* Create session8_insession_guide.rmd

* Rename session7_insession_exercise.rmd to session7_insession_guide.rmd

* Session 8 in-session guide

I created this as generic although I ran a specific experiment with text in the boxes etc. My impression is that's what we're doing because we're facilitating them toward a Q/H/etc. while they learn the mechanics.

* Update session8_insession_guide.rmd

@branscombj I made some minor edits. The activity appears to be done on the individual level but on number 6 the activity says, "Study the system diagram and team data, and decide together what change you want to experiment with, that might give you the desired result." Should there be a note at the beginning that this being done with a group or in a paired format?