Update mtl_live_session12_see.Rmd
Looks like the YAML header doesn't take the R markdown formatting - I took out the asterisks around MTL in the header because it renders as actual asterisks.

I changed the "MTL" in "MTL Live Session 12" to be italicized as well.  Looks like that was the last place that needed it.