Say files for sessions 5, 6 and 7
Ugh - This has taken a lot more time than I expected, and we have much more to do. I chose to work in Word because GitHub/R slows me down so much. so I'm uploading my first 3 Say files as attachments.

One of the challenges is corralling all the great verbiage that has been captured in various places. Would love help with that as we complete these. I know we have some for the variables that oscillate in bc.