Update mtl_session04_say.md (#239)
* Update mtl_session04_say.md

Revisions specific to MTL Video.

* Update mtl_session04_say.md

* Update mtl_session04_say.md

* Update mtl_session06_say.md

Adding more detail to script; adjustments for MTL Video (vs. Live). Putting mousework/directional cues in **bold**.

Not sure if the question about units of time should be included because it's not really answered directly in the model diagram. Actually kind of surprised how few units or hints of units are in the i information in the diagram for AGG... Should that be revisited?

Did not script out "systems story" per se.