Update mtl_session08_see
Not sure it is critical but the gif for the Done doesn't actually show the user clicking to open the Output section and clicking on Expand. Since this step is in the Done text and session 8 is still early in the users learning process for all the parts of the module it may be important to have this showing in a future gif.