+ file upload instruction, - module-specific info
Seemed like there needed to be  a step after 3 that says to copy the filename and a step between 5 & 6 to finish describing how to load the data file. I tried to do as little to change the numbering as possible. If accepted we do have to make the Say file numbers line up.
Could we make the gif about going into the data ui go all the way to pointing to the folder and file?