Update mtl_session07_see.md
One typo.
The only other change viewers might notice from when videos were recorded is the gifs with the Outputs section, which now has Outputs and Text in the blue bar. I personally don't think that warrants redoing the gifs, but defer to HQ. Also we refer in the See guide text to the "Expanded Outputs section" - I think that's fine too.