team_tracker_template.md (#598)
* team_tracker_template.md

Checklist utilized for the team_tracker.

* Update team_tracker_template.md

I didn't have time to review yet in detail so no edits; just these questions:

Could we have *only* fill-in the blanks or questions with a space to answer instead of some checklist items and some fill-in-blanks which sometimes overlap? If we could have that and keep it to, say, 5 items for each session, I feel that would be super helpful and doable.

The checklist does a good job of helping us track our working conceptualization and team tailoring, but do we also need prompts to stay close to the Partner-Build-Apply phases and possibly the CFBT?

Do we plan to look at the team's data *before* Session 2 and be prepared to steer them to look at certain trends? If so, where can we put that? Meta: These are written for post-session completion, I think, and a pre-session, parallel version, would be helpful.