Comparator Intervention Master Brochure
I think we need one brochure for the Data UI comparator intervention, generic to 2 or 12 sessions. I think we'll be following modified scripts for S02 and S03 for the IIR, and doing (sort of) S02 and 11 x S03 for the HSR&D. But I think we can cover all that in one brochure and people can get multiple credit hours when they do the sessions.