Update mtl_session07_say.md
Fixed "Expanded Outputs" language (Outputs and Text) and instructions. Aligned numbering of steps with SEE guide. Removed instructions in the DO about running a new experiment in the team world. They should not be individually doing that, and the DO in the SEE only says log in and review bc to prepare for Exp1.