Update README.md
I am not positive that all of these suggested edits are correct; but hope they help.
I wonder about putting the TeamPSD repositories into a table, with a column for Public/Private and a column for Account (if that's the right word for lzim vs epmo) - so those distinctions would stand out more. Also, duh, columns for name, link, and purpose or description.